requirements-finetune.txt
transformers>=4.45.0,<5.0.0
peft>=0.13.0
trl>=0.12.0,<0.20.0
bitsandbytes>=0.45.0
accelerate>=1.0.0
datasets>=2.14.0
lm-eval>=0.4.0
huggingface-hub>=0.25.0
pyyaml>=6.0
